Revert "refactor: Move code controller to service"
This reverts commit a1c062ffa65d4614a5b5a6849f5444b41725d83d.
