+++ goout-server/src/register/register.controller.ts
@@ -11,12 +11,11 @@ import { RegisterDataDto } from "src/userdata/register.interface";
 import { UserdataService } from "src/userdata/userdata.service";
 import { RegisterAuthNumCheck } from "./registerAuth.interface";
 import * as nodemailer from 'nodemailer'
-import { RegisterService } from "./register.service";
 
 @ApiTags("학생용 라우터")
 @Controller("register")
 export class RegisterController {
-  constructor(private readonly userdataservice: UserdataService,private readonly registerservice:RegisterService) {}
+  constructor(private readonly userdataservice: UserdataService) {}
 
   @ApiOperation({ summary: "회원가입", description: "학생,선생님의 회원가입" })
   @Post()
@@ -29,11 +28,29 @@ export class RegisterController {
         "이메일이 이미 존재합니다",
         HttpStatus.BAD_REQUEST
       );
-      
-        
-    } else {
-      await this.registerservice.send_AuthNum(req)
-      return "이메일 인증 후 로그인이 가능합니다.";
+    }
+    try {
+      let authNum = Math.random().toString().substr(2,6);
+
+      const smtpTransport = nodemailer.createTransport({
+          service: "Gmail",
+          auth: {
+              user: process.env.NODEMAILER_USER,
+              pass: process.env.NODEMAILER_PASS
+            }
+        });
+
+        const mailOptions = {
+            from:process.env.NODEMAILER_USER,
+            to:req.email,
+            subject:"Go-Out 회원가입 E-Mail인증번호",
+            text:`인증번호는 ${authNum}입니다.`
+        };
+        await this.userdataservice.createUserdata(req);
+        return "이메일 인증 후 로그인이 가능합니다.";
+    } catch (error) {
+      console.log(error);
+      throw new HttpException("회원가입 에러", HttpStatus.BAD_REQUEST);
     }
   }
