+++ goout-server/src/register/register.module.ts
@@ -2,10 +2,9 @@ import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { Userdata } from "src/userdata/userdata.entity";
 import { UserdataService } from "src/userdata/userdata.service";
-import { RegisterService } from './register.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Userdata])],
-  providers: [UserdataService, RegisterService],
+  providers: [UserdataService],
 })
 export class RegisterModule {}
