--- goout-server/src/register/register.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RegisterService } from './register.service';
-
-describe('RegisterService', () => {
-  let service: RegisterService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [RegisterService],
-    }).compile();
-
-    service = module.get<RegisterService>(RegisterService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
